+++ demo-ng/app/app.module.ts
@@ -1,0 +1,28 @@
+import { NgModule, NO_ERRORS_SCHEMA } from "@angular/core";
+import { NativeScriptModule } from "nativescript-angular/nativescript.module";
+import { AppRoutingModule } from "./app.routing";
+import { AppComponent } from "./app.component";
+
+import { ItemsComponent } from "./item/items.component";
+
+@NgModule({
+  bootstrap: [
+    AppComponent
+  ],
+  imports: [
+    NativeScriptModule,
+    AppRoutingModule
+  ],
+  declarations: [
+    AppComponent,
+    ItemsComponent
+  ],
+  schemas: [
+    NO_ERRORS_SCHEMA
+  ]
+})
+/*
+Pass your application module to the bootstrapModule function located in main.ts to start your app
+*/
+export class AppModule {
+}